using a service for cross-components
