+++ src/app/app.component.html
@@ -1,3 +1,3 @@
 <app-header />
-<app-user-input (calculate)="onCalculateInvestmentResults($event)"/>
-<app-investment-results [results]="resultsData()"/>
+<app-user-input />
+<app-investment-results />
